refactor: move sendMsgsByWsID to another module
