+++ src/back-ws/room-handler.ts
@@ -1,6 +1,7 @@
 import { roomRepository, userRepository, IUser, gameRepository } from './db/db.js';
 import { Game } from './db/games.js';
-import { ExtendedWebSocket, sendMsgsByWsID } from './websocket-server.js'
+import { sendMsgsByWsID } from './messages/msgs.js';
+import { ExtendedWebSocket } from './websocket-server.js'
 // 2 actions for room: createRoom, addUserToRoom:
 
 export function createRoom(ws: ExtendedWebSocket, data: any, id: number) {
