+++ src/back-ws/websocket-server.ts
@@ -2,10 +2,8 @@ import { randomUUID } from 'node:crypto';
 import { WebSocketServer, WebSocket } from 'ws';
 import { handleRegistration } from './reg-handler.js';
 import { createRoom, addUserToRoom } from './room-handler.js';
-import { wsServer } from '../../index.js'
 import { addShips } from './ships-handler.js';
 import { attack } from './attack-handler.js';
-import { IMsg } from './messages/msgs.js';
 
 export interface ExtendedWebSocket extends WebSocket {
   id: string;
@@ -54,18 +52,3 @@ export const startWebSocketServer = (port: number) => {
 function isMessageHandler(type: any): type is IInputTypeMsg {
   return type in messageHandlers;
 }
-
-export function sendMsgsByWsID(destinationWsIds: string[] | string | 'all', msg: IMsg) {
-  if (typeof destinationWsIds === 'string' &&
-    destinationWsIds !== 'all') { destinationWsIds = [destinationWsIds] }
-
-  const allWsClients = [...wsServer.clients] as ExtendedWebSocket[];
-  let responseList;
-  if (destinationWsIds === 'all') { responseList = allWsClients }
-  else {
-    responseList = allWsClients.filter(client => destinationWsIds.includes(client.id));
-  }
-
-  const message = JSON.stringify(msg);
-  responseList.forEach(client => client.send(message));
-};
